requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-extensions==1.9.0
django-simple-history==1.9.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
idna==2.6
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
PyJWT==1.5.3
PyMySQL==0.7.11
pytz==2017.2
requests==2.18.4
six==1.10.0
suds-py3==1.3.3.0
urllib3==1.22
virtualenv==15.1.0
